fix(module:message): fix lazy load problem
close #3794
