+++ components/core/public-api.ts
@@ -23,3 +23,4 @@ export * from './dropdown/public-api';
 export * from './logger/public-api';
 export * from './responsive/public-api';
 export * from './trans-button/public-api';
+export * from './message/public-api';
