+++ components/message/nz-message.service.ts
@@ -7,87 +7,17 @@
  */
 
 import { Overlay } from '@angular/cdk/overlay';
-import {
-  ApplicationRef,
-  ComponentFactoryResolver,
-  EmbeddedViewRef,
-  Injectable,
-  Injector,
-  TemplateRef,
-  Type
-} from '@angular/core';
+import { ApplicationRef, ComponentFactoryResolver, Injectable, Injector, TemplateRef } from '@angular/core';
+
+import { NzMessageBaseService } from 'ng-zorro-antd/core';
 
 import { NzMessageConfig } from './nz-message-config';
 import { NzMessageContainerComponent } from './nz-message-container.component';
 import { NzMessageData, NzMessageDataFilled, NzMessageDataOptions } from './nz-message.definitions';
-
-let globalCounter = 0;
-
-export class NzMessageBaseService<
-  ContainerClass extends NzMessageContainerComponent,
-  MessageData,
-  MessageConfig extends NzMessageConfig
-> {
-  protected _container: ContainerClass;
-
-  constructor(
-    private overlay: Overlay,
-    private containerClass: Type<ContainerClass>,
-    private injector: Injector,
-    private cfr: ComponentFactoryResolver,
-    private appRef: ApplicationRef,
-    private _idPrefix: string = ''
-  ) {
-    this._container = this.createContainer();
-  }
-
-  remove(messageId?: string): void {
-    if (messageId) {
-      this._container.removeMessage(messageId);
-    } else {
-      this._container.removeMessageAll();
-    }
-  }
-
-  createMessage(message: MessageData, options?: NzMessageDataOptions): NzMessageDataFilled {
-    const resultMessage: NzMessageDataFilled = {
-      ...(message as NzMessageData),
-      ...{
-        createdAt: new Date(),
-        messageId: this._generateMessageId(),
-        options
-      }
-    };
-    this._container.createMessage(resultMessage);
-
-    return resultMessage;
-  }
-
-  config(config: MessageConfig): void {
-    this._container.setConfig(config);
-  }
-
-  protected _generateMessageId(): string {
-    return this._idPrefix + globalCounter++;
-  }
-
-  // Manually creating container for overlay to avoid multi-checking error, see: https://github.com/NG-ZORRO/ng-zorro-antd/issues/391
-  // NOTE: we never clean up the container component and it's overlay resources, if we should, we need to do it by our own codes.
-  private createContainer(): ContainerClass {
-    const factory = this.cfr.resolveComponentFactory(this.containerClass);
-    const componentRef = factory.create(this.injector); // Use root injector
-    componentRef.changeDetectorRef.detectChanges(); // Immediately change detection to avoid multi-checking error
-    this.appRef.attachView(componentRef.hostView); // Load view into app root
-    const overlayPane = this.overlay.create().overlayElement;
-    overlayPane.style.zIndex = '1010'; // Patching: assign the same zIndex of ant-message to it's parent overlay panel, to the ant-message's zindex work.
-    overlayPane.appendChild((componentRef.hostView as EmbeddedViewRef<{}>).rootNodes[0] as HTMLElement);
-
-    return componentRef.instance;
-  }
-}
+import { NzMessageServiceModule } from './nz-message.service.module';
 
 @Injectable({
-  providedIn: 'root'
+  providedIn: NzMessageServiceModule
 })
 export class NzMessageService extends NzMessageBaseService<
   NzMessageContainerComponent,
