+++ components/message/public-api.ts
@@ -7,6 +7,7 @@
  */
 
 export * from './nz-message.service';
+export * from './nz-message.service.module';
 export * from './nz-message.module';
 export * from './nz-message.component';
 export * from './nz-message.definitions';
